refactor: split ui out
