+++ sim/index.html
@@ -14,7 +14,7 @@
     Your browser doesn't support HTML canvas.
   </canvas>
 
-  <!-- Trunk will compile `src/ui.rs` into Wasm & JS and inject them here. -->
-  <script data-trunk src="src/ui.rs"></script>
+  <!-- Trunk will compile `src/ui/core.rs` into Wasm & JS and inject them here. -->
+  <script data-trunk src="src/ui/core.rs"></script>
 </body>
 </html>
