--- sim/src/ui.rs
@@ -1,290 +1,0 @@
-use js_sys::Date;
-use std::cell::RefCell;
-use wasm_bindgen::prelude::*;
-use wasm_bindgen::{JsCast, closure::Closure};
-use web_sys::{
-    CanvasRenderingContext2d, Document, HtmlCanvasElement, MouseEvent, WheelEvent, console,
-};
-
-use crate::model::{GLOBAL_STATE, VehicleType};
-
-/// Stores camera offset (pan_x, pan_y) and a zoom factor (`scale`).
-#[derive(Default)]
-struct Camera {
-    pub pan_x: f32,
-    pub pan_y: f32,
-    pub scale: f32,
-}
-
-/// Whether user is dragging, plus last mouse coords
-#[derive(Default)]
-struct DragState {
-    pub is_dragging: bool,
-    pub last_x: f32,
-    pub last_y: f32,
-}
-
-// A thread-local "camera" storing our pan offset and zoom scale.
-thread_local! {
-    static CAMERA: RefCell<Camera> = RefCell::new(Camera { pan_x: 0.0, pan_y: 0.0, scale: 1.0 });
-    static DRAG: RefCell<DragState> = RefCell::new(DragState::default());
-}
-
-/// Called once when the Wasm module loads:
-/// 1) Initialize routes/vehicles
-/// 2) Start the update loop (setInterval ~60 FPS)
-/// 3) Attach mouse events for panning and wheel event for zoom
-#[wasm_bindgen(start)]
-pub fn main_js() -> Result<(), JsValue> {
-    console_error_panic_hook::set_once();
-
-    // 1) Create routes & vehicles (random or real)
-    GLOBAL_STATE.with(|cell| {
-        let mut state = cell.borrow_mut();
-        state.init_vehicles();
-    });
-
-    // 2) Repeated update & draw
-    start_animation_loop()?;
-
-    // 3) Attach mouse & wheel listeners
-    attach_mouse_listeners()?;
-    attach_wheel_listener()?;
-
-    Ok(())
-}
-
-/// Creates a closure that runs at ~60 FPS (setInterval) and updates + draws.
-fn start_animation_loop() -> Result<(), JsValue> {
-    let closure = Closure::wrap(Box::new(move || {
-        let t_start = Date::now();
-
-        // 1) Update vehicles
-        GLOBAL_STATE.with(|cell| {
-            cell.borrow_mut().update_all();
-        });
-
-        // 2) Draw everything
-        if let Some(window) = web_sys::window() {
-            if let Some(document) = window.document() {
-                if let Some(canvas_el) = document.get_element_by_id("myCanvas") {
-                    if let Ok(canvas) = canvas_el.dyn_into::<HtmlCanvasElement>() {
-                        if let Ok(ctx) = canvas
-                            .get_context("2d")
-                            .unwrap()
-                            .unwrap()
-                            .dyn_into::<CanvasRenderingContext2d>()
-                        {
-                            let w = canvas.width() as f64;
-                            let h = canvas.height() as f64;
-                            ctx.clear_rect(0.0, 0.0, w, h);
-
-                            draw_routes(&ctx);
-                            draw_vehicles(&ctx);
-                        }
-                    }
-                }
-            }
-        }
-
-        let t_end = Date::now();
-        let ms = t_end - t_start;
-        console::log_1(&format!("Update & draw took {:.3} ms", ms).into());
-    }) as Box<dyn FnMut()>);
-
-    // setInterval(..., 16) => ~60 FPS
-    let window = web_sys::window().unwrap();
-    window.set_interval_with_callback_and_timeout_and_arguments_0(
-        closure.as_ref().unchecked_ref(),
-        16,
-    )?;
-
-    // Keep the closure alive
-    closure.forget();
-
-    Ok(())
-}
-
-/// Draw routes, offsetting by the camera's pan_x/pan_y and scaling by camera.scale
-fn draw_routes(ctx: &CanvasRenderingContext2d) {
-    let (pan_x, pan_y, scale) = CAMERA.with(|c| {
-        let cam = c.borrow();
-        (cam.pan_x, cam.pan_y, cam.scale)
-    });
-
-    GLOBAL_STATE.with(|cell| {
-        let state = cell.borrow();
-        for (i, route) in state.routes.iter().enumerate() {
-            // color routes differently for trains vs buses
-            if i < 10 {
-                ctx.set_fill_style_str("rgba(255, 150, 150, 0.6)");
-            } else {
-                ctx.set_fill_style_str("rgba(150, 150, 255, 0.6)");
-            }
-
-            for &(sx, sy) in &route.stations {
-                ctx.begin_path();
-
-                // shift by pan_x / pan_y, then apply scale
-                // final_x = (sx - pan_x)*scale
-                let draw_x = (sx - pan_x) * scale;
-                let draw_y = (sy - pan_y) * scale;
-
-                ctx.arc(draw_x as f64, draw_y as f64, 5.0 * scale as f64, 0.0, 6.28)
-                    .unwrap();
-                ctx.fill();
-            }
-        }
-    });
-}
-
-/// Draw vehicles, offset by camera pan and scale
-fn draw_vehicles(ctx: &CanvasRenderingContext2d) {
-    let (pan_x, pan_y, scale) = CAMERA.with(|c| {
-        let cam = c.borrow();
-        (cam.pan_x, cam.pan_y, cam.scale)
-    });
-
-    GLOBAL_STATE.with(|cell| {
-        let state = cell.borrow();
-        for v in &state.vehicles {
-            ctx.begin_path();
-
-            let draw_x = (v.x - pan_x) * scale;
-            let draw_y = (v.y - pan_y) * scale;
-
-            // vehicles are small, so also multiply the circle radius by scale
-            ctx.arc(draw_x as f64, draw_y as f64, 2.0 * scale as f64, 0.0, 6.28)
-                .unwrap();
-
-            match v.vehicle_type {
-                VehicleType::Bus => ctx.set_fill_style_str("blue"),
-                VehicleType::Train => ctx.set_fill_style_str("red"),
-            }
-            ctx.fill();
-        }
-    });
-}
-
-/// Attach mouse events so user can pan the canvas.
-fn attach_mouse_listeners() -> Result<(), JsValue> {
-    let window = web_sys::window().unwrap();
-    let document: Document = window.document().unwrap();
-    let canvas_el = document
-        .get_element_by_id("myCanvas")
-        .ok_or("no #myCanvas element")?;
-
-    // MOUSE DOWN
-    {
-        let closure_mousedown = Closure::wrap(Box::new(move |e: MouseEvent| {
-            DRAG.with(|ds| {
-                let mut drag = ds.borrow_mut();
-                drag.is_dragging = true;
-                drag.last_x = e.client_x() as f32;
-                drag.last_y = e.client_y() as f32;
-            });
-        }) as Box<dyn FnMut(_)>);
-
-        canvas_el.add_event_listener_with_callback(
-            "mousedown",
-            closure_mousedown.as_ref().unchecked_ref(),
-        )?;
-        closure_mousedown.forget();
-    }
-
-    // MOUSE UP
-    {
-        let closure_mouseup = Closure::wrap(Box::new(move |_e: MouseEvent| {
-            DRAG.with(|ds| {
-                ds.borrow_mut().is_dragging = false;
-            });
-        }) as Box<dyn FnMut(_)>);
-
-        // We can add mouseup to the canvas or the entire document
-        document.add_event_listener_with_callback(
-            "mouseup",
-            closure_mouseup.as_ref().unchecked_ref(),
-        )?;
-        closure_mouseup.forget();
-    }
-
-    // MOUSE MOVE
-    {
-        let closure_mousemove = Closure::wrap(Box::new(move |e: MouseEvent| {
-            let new_x = e.client_x() as f32;
-            let new_y = e.client_y() as f32;
-
-            DRAG.with(|ds| {
-                let mut drag = ds.borrow_mut();
-                if drag.is_dragging {
-                    // compute delta
-                    let dx = new_x - drag.last_x;
-                    let dy = new_y - drag.last_y;
-
-                    // shift camera
-                    CAMERA.with(|c| {
-                        let mut cam = c.borrow_mut();
-                        // dragging right => negative offset
-                        cam.pan_x -= dx / cam.scale; // incorporate scale if you like
-                        cam.pan_y -= dy / cam.scale;
-                    });
-
-                    // update last
-                    drag.last_x = new_x;
-                    drag.last_y = new_y;
-                }
-            });
-        }) as Box<dyn FnMut(_)>);
-
-        document.add_event_listener_with_callback(
-            "mousemove",
-            closure_mousemove.as_ref().unchecked_ref(),
-        )?;
-        closure_mousemove.forget();
-    }
-
-    Ok(())
-}
-
-/// Attach a wheel event to allow zoom in/out with the mouse wheel.
-fn attach_wheel_listener() -> Result<(), JsValue> {
-    let window = web_sys::window().unwrap();
-    let document: Document = window.document().unwrap();
-    let canvas_el = document
-        .get_element_by_id("myCanvas")
-        .ok_or("no #myCanvas element")?;
-
-    // We'll do a quick "scale *= factor" approach.  Use SHIFT or trackpad direction
-    // to pick the factor. e.g. factor = 1.1 => 10% zoom in, or 0.9 => 10% zoom out
-    let closure_wheel = Closure::wrap(Box::new(move |e: WheelEvent| {
-        e.prevent_default(); // prevent page scroll
-        let delta = e.delta_y(); // positive => zoom out, negative => zoom in
-        CAMERA.with(|c| {
-            let mut cam = c.borrow_mut();
-
-            if delta < 0.0 {
-                // zoom in
-                cam.scale *= 1.1;
-            } else {
-                // zoom out
-                cam.scale *= 0.9;
-            }
-
-            // clamp scale
-            if cam.scale < 0.1 {
-                cam.scale = 0.1;
-            } else if cam.scale > 50.0 {
-                cam.scale = 50.0;
-            }
-
-            // Optional: we can also keep the mouse position stable
-            // by offsetting pan_x/pan_y so that the point under the mouse
-            // remains the same after zoom. That's more advanced, see below.
-        });
-    }) as Box<dyn FnMut(_)>);
-
-    canvas_el.add_event_listener_with_callback("wheel", closure_wheel.as_ref().unchecked_ref())?;
-    closure_wheel.forget();
-
-    Ok(())
-}
