+++ sim/src/ui/mod.rs
@@ -1,0 +1,6 @@
+// src/ui/mod.rs
+
+pub mod core;
+pub mod camera;
+pub mod draw;
+pub mod input;
